requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
requests
httpx